# perf(core/examples): optimize single writes