```
# main

Running 10s test @ http://localhost:4500
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    56.90us   25.87us   2.23ms   85.59%
    Req/Sec    78.31k     2.36k   80.80k    98.02%
  Latency Distribution
     50%   56.00us
     75%   66.00us
     90%   78.00us
     99%  106.00us
  1573901 requests in 10.10s, 76.55MB read
Requests/sec: 155834.47
Transfer/sec:      7.58MB
```

```
# This patch

Running 10s test @ http://localhost:4544
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    51.05us   16.60us 380.00us   71.85%
    Req/Sec    86.00k     4.06k   97.58k    90.22%
  Latency Distribution
     50%   49.00us
     75%   60.00us
     90%   72.00us
     99%   97.00us
  1575430 requests in 9.20s, 76.62MB read
Requests/sec: 171248.63
Transfer/sec:      8.33MB
```